DRY: Extract shared batchSync<T>() helper from sync scripts
Extracts ~550 lines of duplicated batch sync logic from 7 sync scripts
into a generic batchSync<T>() function in sync-common.ts.

The new helper handles: batching, POST with retry, consecutive failure
detection, progress logging, and error counting. Each sync script now
defines only its data loading/transformation and calls batchSync() with
its endpoint URL, body key, and response count key.

Also moves waitForHealthy() and fetchWithRetry() from sync-pages.ts to
sync-common.ts (re-exported from sync-pages.ts for backward compat).

All 98 existing tests pass unchanged.

Closes #504
