+++ crux/wiki-server/sync-edit-logs.ts
@@ -4,7 +4,7 @@
  * Reads all data/edit-logs/*.yaml files and bulk-upserts them to the
  * wiki-server's /api/edit-logs/batch endpoint.
  *
- * Reuses the retry + health-check pattern from sync-pages.ts.
+ * Reuses the shared batch sync infrastructure from sync-common.ts.
  *
  * Usage:
  *   pnpm crux wiki-server sync-edit-logs
@@ -21,15 +21,14 @@ import { join } from 'path';
 import { fileURLToPath } from 'url';
 import { parse as parseYaml } from 'yaml';
 import { parseCliArgs } from '../lib/cli.ts';
-import { getServerUrl, getApiKey, buildHeaders, type EditLogApiEntry } from '../lib/wiki-server-client.ts';
-import { waitForHealthy, fetchWithRetry } from './sync-pages.ts';
+import { getServerUrl, getApiKey, type EditLogApiEntry } from '../lib/wiki-server-client.ts';
+import { waitForHealthy, batchSync } from './sync-common.ts';
 
 const PROJECT_ROOT = join(import.meta.dirname!, '../..');
 const EDIT_LOGS_DIR = join(PROJECT_ROOT, 'data/edit-logs');
 
 // --- Configuration ---
 const DEFAULT_BATCH_SIZE = 200;
-const MAX_CONSECUTIVE_FAILURES = 3;
 
 // --- Types ---
 
@@ -108,64 +107,19 @@ export async function syncEditLogs(
     _sleep?: (ms: number) => Promise<void>;
   } = {},
 ): Promise<{ inserted: number; errors: number }> {
-  let totalInserted = 0;
-  let totalErrors = 0;
-  let consecutiveFailures = 0;
+  const result = await batchSync(
+    `${serverUrl}/api/edit-logs/batch`,
+    items,
+    batchSize,
+    {
+      bodyKey: 'items',
+      responseCountKey: 'inserted',
+      itemLabel: 'entries',
+      _sleep: options._sleep,
+    },
+  );
 
-  for (let i = 0; i < items.length; i += batchSize) {
-    const batch = items.slice(i, i + batchSize);
-    const batchNum = Math.floor(i / batchSize) + 1;
-    const totalBatches = Math.ceil(items.length / batchSize);
-
-    try {
-      const res = await fetchWithRetry(
-        `${serverUrl}/api/edit-logs/batch`,
-        {
-          method: 'POST',
-          headers: buildHeaders(),
-          body: JSON.stringify({ items: batch }),
-        },
-        { _sleep: options._sleep },
-      );
-
-      if (!res.ok) {
-        const body = await res.text();
-        console.error(
-          `  Batch ${batchNum}/${totalBatches}: HTTP ${res.status} — ${body}`,
-        );
-        totalErrors += batch.length;
-        consecutiveFailures++;
-      } else {
-        const result = (await res.json()) as { inserted: number };
-        totalInserted += result.inserted;
-        consecutiveFailures = 0;
-
-        console.log(
-          `  Batch ${batchNum}/${totalBatches}: ${result.inserted} inserted`,
-        );
-      }
-    } catch (err) {
-      console.error(
-        `  Batch ${batchNum}/${totalBatches}: Failed after retries — ${err}`,
-      );
-      totalErrors += batch.length;
-      consecutiveFailures++;
-    }
-
-    if (consecutiveFailures >= MAX_CONSECUTIVE_FAILURES) {
-      const remaining = items.length - (i + batchSize);
-      if (remaining > 0) {
-        console.error(
-          `\n  Aborting: ${MAX_CONSECUTIVE_FAILURES} consecutive batch failures. ` +
-            `Skipping ${remaining} remaining entries.`,
-        );
-        totalErrors += remaining;
-      }
-      break;
-    }
-  }
-
-  return { inserted: totalInserted, errors: totalErrors };
+  return { inserted: result.count, errors: result.errors };
 }
 
 // --- CLI ---
