+++ crux/wiki-server/sync-resources.ts
@@ -4,7 +4,7 @@
  * Reads all data/resources/*.yaml files and bulk-upserts them to the
  * wiki-server's /api/resources/batch endpoint.
  *
- * Reuses the retry + health-check pattern from sync-pages.ts:
+ * Reuses the shared batch sync infrastructure from sync-common.ts:
  *   - Pre-sync health check with retries (waits for server to be ready)
  *   - Per-batch retry with exponential backoff (handles transient 5xx errors)
  *   - Fast-fail after N consecutive batch failures
@@ -24,15 +24,14 @@ import { join } from "path";
 import { fileURLToPath } from "url";
 import { parse as parseYaml } from "yaml";
 import { parseCliArgs } from "../lib/cli.ts";
-import { getServerUrl, getApiKey, buildHeaders } from "../lib/wiki-server-client.ts";
-import { waitForHealthy, fetchWithRetry } from "./sync-pages.ts";
+import { getServerUrl, getApiKey } from "../lib/wiki-server-client.ts";
+import { waitForHealthy, batchSync } from "./sync-common.ts";
 
 const PROJECT_ROOT = join(import.meta.dirname!, "../..");
 const RESOURCES_DIR = join(PROJECT_ROOT, "data/resources");
 
 // --- Configuration ---
 const DEFAULT_BATCH_SIZE = 100;
-const MAX_CONSECUTIVE_FAILURES = 3;
 
 // --- Types ---
 
@@ -178,64 +177,19 @@ export async function syncResources(
     _sleep?: (ms: number) => Promise<void>;
   } = {}
 ): Promise<{ upserted: number; errors: number }> {
-  let totalUpserted = 0;
-  let totalErrors = 0;
-  let consecutiveFailures = 0;
+  const result = await batchSync(
+    `${serverUrl}/api/resources/batch`,
+    items,
+    batchSize,
+    {
+      bodyKey: "items",
+      responseCountKey: "inserted",
+      itemLabel: "resources",
+      _sleep: options._sleep,
+    },
+  );
 
-  for (let i = 0; i < items.length; i += batchSize) {
-    const batch = items.slice(i, i + batchSize);
-    const batchNum = Math.floor(i / batchSize) + 1;
-    const totalBatches = Math.ceil(items.length / batchSize);
-
-    try {
-      const res = await fetchWithRetry(
-        `${serverUrl}/api/resources/batch`,
-        {
-          method: "POST",
-          headers: buildHeaders(),
-          body: JSON.stringify({ items: batch }),
-        },
-        { _sleep: options._sleep }
-      );
-
-      if (!res.ok) {
-        const body = await res.text();
-        console.error(
-          `  Batch ${batchNum}/${totalBatches}: HTTP ${res.status} — ${body}`
-        );
-        totalErrors += batch.length;
-        consecutiveFailures++;
-      } else {
-        const result = (await res.json()) as { inserted: number };
-        totalUpserted += result.inserted;
-        consecutiveFailures = 0;
-
-        console.log(
-          `  Batch ${batchNum}/${totalBatches}: ${result.inserted} upserted`
-        );
-      }
-    } catch (err) {
-      console.error(
-        `  Batch ${batchNum}/${totalBatches}: Failed after retries — ${err}`
-      );
-      totalErrors += batch.length;
-      consecutiveFailures++;
-    }
-
-    if (consecutiveFailures >= MAX_CONSECUTIVE_FAILURES) {
-      const remaining = items.length - (i + batchSize);
-      if (remaining > 0) {
-        console.error(
-          `\n  Aborting: ${MAX_CONSECUTIVE_FAILURES} consecutive batch failures. ` +
-            `Skipping ${remaining} remaining resources.`
-        );
-        totalErrors += remaining;
-      }
-      break;
-    }
-  }
-
-  return { upserted: totalUpserted, errors: totalErrors };
+  return { upserted: result.count, errors: result.errors };
 }
 
 // --- CLI ---
